fix(issue):Issue#1
Don't expose secrets in plaintext, nor git versioning
