+++ features/auth/data/src/main/java/com/kyawlinnthant/data/repository/AuthApiRepository.kt
@@ -1,13 +1,13 @@
 package com.kyawlinnthant.data.repository
 
 import com.kyawlinnthant.data.model.ForgotPasswordResponse
+import com.kyawlinnthant.data.model.LoginRequest
 import com.kyawlinnthant.data.model.LoginResponse
 import com.kyawlinnthant.network.util.DataResult
 
 interface AuthApiRepository {
     suspend fun login(
-        email: String,
-        password: String
+        body: LoginRequest
     ): DataResult<LoginResponse>
 
     suspend fun forgotPassword(
