+++ features/auth/data/src/main/java/com/kyawlinnthant/data/repository/AuthApiRepositoryImpl.kt
@@ -1,6 +1,5 @@
 package com.kyawlinnthant.data.repository
 
-import com.kyawlinnthant.auth.data.BuildConfig
 import com.kyawlinnthant.data.model.ForgotPasswordResponse
 import com.kyawlinnthant.data.model.LoginRequest
 import com.kyawlinnthant.data.model.LoginResponse
@@ -16,14 +15,9 @@ class AuthApiRepositoryImpl @Inject constructor(
     private val api: AuthService,
     @DispatcherModule.IoDispatcher private val io: CoroutineDispatcher
 ) : AuthApiRepository {
-    override suspend fun login(email: String, password: String): DataResult<LoginResponse> {
-        val body = LoginRequest(
-            type = "password",
-            email = email,
-            password = password,
-            clientId = BuildConfig.CLIENT_ID,
-            clientSecret = BuildConfig.CLIENT_SECRET
-        )
+    override suspend fun login(
+        body: LoginRequest
+    ): DataResult<LoginResponse> {
         return withContext(io) {
             safeApiCall {
                 api.login(body = body)
